+terms and converters, various code, naming and docs improvements, airways close to final

+++ source/terms/BoundaryVia.cs
@@ -1,13 +1,10 @@
-using Arinc.Spec424.Attributes;
-using Arinc.Spec424.Terms.Converters;
-
 namespace Arinc.Spec424.Terms;
 
 /// <summary>
-/// <c>Boundary Via (BDRY VIA)</c> character. See paragraph 5.118.
+/// <c>Boundary Via (BDRY VIA)</c> character.
 /// </summary>
-/// <remarks><see cref="TransformAttribute">Transformed</see> by <see cref="BoundaryViaConverter"/>.</remarks>
-public enum BoundaryVia : int
+/// <remarks>See section 5.118.</remarks>
+public enum BoundaryVia : byte
 {
     Unknown,
     Circle,
